+++ deepparse/tools.py
@@ -1,9 +1,12 @@
 import os
+import shutil
 import warnings
 from typing import List
 
 import poutyne
 import requests
+from requests import HTTPError
+from urllib3.exceptions import MaxRetryError
 
 from .data_error import DataError
 from .data_validation import (
@@ -15,17 +18,67 @@ from .data_validation import (
 BASE_URL = "https://graal.ift.ulaval.ca/public/deepparse/{}.{}"
 CACHE_PATH = os.path.join(os.path.expanduser("~"), ".cache", "deepparse")
 
+# Status code starting in the 4xx are client error status code.
+# That is Deepparse, server problem (e.g. Deepparse server is offline).
+HTTP_CLIENT_ERROR_STATUS_CODE = 400
+# Status code starting in the 5xx are the next range status code.
+NEXT_RANGE_STATUS_CODE = 500
 
-def latest_version(model: str, cache_path: str) -> bool:
+
+def latest_version(model: str, cache_path: str, verbose: bool) -> bool:
     """
     Verify if the local model is the latest.
     """
+    # Reading of the actual local version
     with open(os.path.join(cache_path, model + ".version"), encoding="utf-8") as local_model_hash_file:
         local_model_hash_version = local_model_hash_file.readline()
-    download_from_url(model, cache_path, "version")
-    with open(os.path.join(cache_path, model + ".version"), encoding="utf-8") as remote_model_hash_file:
-        remote_model_hash_version = remote_model_hash_file.readline()
-    return local_model_hash_version.strip() == remote_model_hash_version.strip()
+    try:
+        # We create a temporary directory for the server-side version file
+        tmp_cache = os.path.join(cache_path, "tmp")
+        os.makedirs(tmp_cache, exist_ok=True)
+
+        download_from_url(model, tmp_cache, "version")
+
+        # Reading of the server-side version
+        with open(os.path.join(tmp_cache, model + ".version"), encoding="utf-8") as remote_model_hash_file:
+            remote_model_hash_version = remote_model_hash_file.readline()
+
+        is_latest_version = local_model_hash_version.strip() == remote_model_hash_version.strip()
+
+    except HTTPError as exception:  # HTTP connection error handling
+        if HTTP_CLIENT_ERROR_STATUS_CODE <= exception.response.status_code < NEXT_RANGE_STATUS_CODE:
+            # Case where Deepparse server is down.
+            if verbose:
+                warnings.warn(
+                    f"We where not able to verify the cached model in the cache directory {cache_path}. It seems like"
+                    f"Deepparse server is not available at the moment. We recommend to attempt to verify "
+                    f"the model version another time using our download CLI function."
+                )
+            # The is_lastest_version is set to True even if we were not able to validate the version. We do so not to
+            # block the rest of the process.
+            is_latest_version = True
+        else:
+            # We re-raise the exception if the status_code is not in the two ranges we are interested in
+            # (local server or remote server error).
+            raise
+    except MaxRetryError:
+        # Case where the user does not have an Internet connection. For example, one can run it in a
+        # Docker container not connected to the Internet.
+        if verbose:
+            warnings.warn(
+                f"We where not able to verify the cached model in the cache directory {cache_path}. It seems like"
+                f"you are not connected to the Internet. We recommend to verify if you have the latest using our "
+                f"download CLI function."
+            )
+        # The is_lastest_version is set to True even if we were not able to validate the version. We do so not to
+        # block the rest of the process.
+        is_latest_version = True
+    finally:
+        # Cleaning the temporary directory
+        if os.path.exists(tmp_cache):
+            shutil.rmtree(tmp_cache)
+
+    return is_latest_version
 
 
 def download_from_url(file_name: str, saving_dir: str, file_extension: str):
@@ -85,48 +138,6 @@ def valid_poutyne_version(min_major: int = 1, min_minor: int = 2):
         is_valid_poutyne_version = major >= min_major and minor >= min_minor
 
     return is_valid_poutyne_version
-
-
-def handle_pre_trained_checkpoint(model_type_checkpoint: str) -> str:
-    """
-    Handle the checkpoint formatting for pretrained models.
-    """
-    if not valid_poutyne_version():
-        raise NotImplementedError(
-            f"To load the pretrained {model_type_checkpoint} model, you need to have a Poutyne version"
-            "greater than 1.1 (>1.1)"
-        )
-    model_path = os.path.join(CACHE_PATH, f"{model_type_checkpoint}.ckpt")
-
-    if not os.path.isfile(model_path):
-        download_weights(model_type_checkpoint, CACHE_PATH, verbose=True)
-    elif not latest_version(model_type_checkpoint, cache_path=CACHE_PATH):
-        warnings.warn(
-            "A newer model of fasttext is available, you can download it using the download script.",
-            UserWarning,
-        )
-    checkpoint = os.path.join(CACHE_PATH, f"{model_type_checkpoint}.ckpt")
-    return checkpoint
-
-
-def handle_model_path(checkpoint: str) -> str:
-    """
-    Handle the validity of path.
-    """
-    if checkpoint in ("fasttext", "bpemb"):
-        checkpoint = handle_pre_trained_checkpoint(checkpoint)
-    elif isinstance(checkpoint, str) and checkpoint.endswith(".ckpt"):
-        if not valid_poutyne_version():
-            raise NotImplementedError(
-                "To load a string path to a model, you need to have a Poutyne version" "greater than 1.1 (>1.1)"
-            )
-    else:
-        raise ValueError(
-            "The checkpoint is not valid. Can be a path in a string format (e.g. 'a_path_.ckpt'), "
-            "'fasttext' or 'bpemb'."
-        )
-
-    return checkpoint
 
 
 def validate_data_to_parse(addresses_to_parse: List) -> None: