+++ tests/cli/test_retrain.py
@@ -5,14 +5,17 @@ import unittest
 from tempfile import TemporaryDirectory
 from typing import List
 from unittest import skipIf
-
-import torch
+from unittest.mock import patch
 
 from deepparse.cli import retrain
 from deepparse.cli.retrain import get_args, parse_retrained_arguments
 from tests.parser.integration.base_retrain import RetrainTestCase
 
 
+@skipIf(
+    not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
+    "download of model too long for test in runner",
+)
 class RetrainTests(RetrainTestCase):
     @classmethod
     def setUpClass(cls):
@@ -29,6 +32,8 @@ class RetrainTests(RetrainTestCase):
 
         cls.a_named_model = "a_retrained_model"
 
+        cls.a_cache_dir = "a_cache_dir"
+
     def setUp(self) -> None:
         self.temp_checkpoints_obj = TemporaryDirectory()
 
@@ -38,31 +43,41 @@ class RetrainTests(RetrainTestCase):
         # temp directory
         self.logging_path = os.path.join(self.temp_checkpoints_obj.name, "checkpoints")
 
-        # We set a set of defaults parser argument to adjust it for the tests
-        # We change the default logging path to the one created as a temp
-        # directory for cleanup, and we reduce the number of epoch tho reduce
-        # test duration.
-        self.parser_test_default_settings = ["--logging_path", self.logging_path, "--epochs", "1"]
-
     def tearDown(self) -> None:
         self.temp_checkpoints_obj.cleanup()
 
     def set_up_params(
         self,
+        model_type=None,  # None to handle the default tests case.
+        train_dataset_path=None,  # None to handle the default tests case.
         train_ratio="0.8",
         batch_size="32",
-        epochs="5",
+        epochs="1",  # As opposed to default CLI function, we set the epoch value number to 1,
         num_workers="1",
         learning_rate="0.01",
         seed="42",
-        logging_path="./checkpoints",
+        logging_path=None,  # None to handle the default case logging path for the tests.
         disable_tensorboard="False",
         layers_to_freeze='seq2seq',
-        name_of_the_retrain_parser="",
+        name_of_the_retrain_parser=None,
+        cache_dir=None,  # None to handle the default library default value (CACHE_PATH)
+        device="cpu",  # By default, we set it to cpu instead of gpu device 0 as the CLI function.
+        csv_column_names: List = None,
+        csv_column_separator="\t",
     ) -> List:
+        if model_type is None:
+            # The default case for the test is a FastText model
+            model_type = self.a_fasttext_model_type
+
+        if train_dataset_path is None:
+            train_dataset_path = self.a_train_pickle_dataset_path
+
+        if logging_path is None:
+            logging_path = self.logging_path
+
         parser_params = [
-            self.a_fasttext_model_type,
-            self.a_train_pickle_dataset_path,
+            model_type,
+            train_dataset_path,
             "--train_ratio",
             train_ratio,
             "--batch_size",
@@ -81,23 +96,28 @@ class RetrainTests(RetrainTestCase):
             disable_tensorboard,
             "--layers_to_freeze",
             layers_to_freeze,
-            "--name_of_the_retrain_parser",
-            name_of_the_retrain_parser,
+            "--device",
+            device,
+            "--csv_column_separator",
+            csv_column_separator,
         ]
+
+        if cache_dir is not None:
+            # To handle the None case (that is using the default None of the argparser).
+            parser_params.extend(["--cache_dir", cache_dir])
+
+        if name_of_the_retrain_parser is not None:
+            # To handle the None case (that is using the default None of the argparser).
+            parser_params.extend(["--name_of_the_retrain_parser", name_of_the_retrain_parser])
+
+        if csv_column_names is not None:
+            parser_params.extend(["--csv_column_names"])
+            parser_params.extend(csv_column_names)  # Since csv_column_names is a list
+
         return parser_params
 
-    @skipIf(
-        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
-        "download of model too long for test in runner",
-    )
     def test_integration_cpu(self):
-        parser_params = [
-            self.a_fasttext_model_type,
-            self.a_train_pickle_dataset_path,
-            "--device",
-            self.cpu_device,
-        ]
-        parser_params.extend(self.parser_test_default_settings)
+        parser_params = self.set_up_params(device=self.cpu_device)
         retrain.main(parser_params)
 
         self.assertTrue(
@@ -106,16 +126,8 @@ class RetrainTests(RetrainTestCase):
             )
         )
 
-    @skipIf(not torch.cuda.is_available(), "no gpu available")
     def test_integration_gpu(self):
-        parser_params = [
-            self.a_fasttext_model_type,
-            self.a_train_pickle_dataset_path,
-            "--device",
-            self.gpu_device,
-        ]
-
-        parser_params.extend(self.parser_test_default_settings)
+        parser_params = self.set_up_params(device=self.gpu_device)
         retrain.main(parser_params)
 
         self.assertTrue(
@@ -124,16 +136,8 @@ class RetrainTests(RetrainTestCase):
             )
         )
 
-    @skipIf(not torch.cuda.is_available(), "no gpu available")
     def test_integration_attention_model(self):
-        parser_params = [
-            self.a_fasttext_att_model_type,
-            self.a_train_pickle_dataset_path,
-            "--device",
-            self.cpu_device,
-        ]
-
-        parser_params.extend(self.parser_test_default_settings)
+        parser_params = self.set_up_params(model_type=self.a_fasttext_att_model_type)
         retrain.main(parser_params)
 
         self.assertTrue(
@@ -144,24 +148,9 @@ class RetrainTests(RetrainTestCase):
             )
         )
 
-    @skipIf(
-        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
-        "download of model too long for test in runner",
-    )
     def test_integration_csv(self):
-        parser_params = [
-            self.a_fasttext_model_type,
-            self.a_train_csv_dataset_path,
-            "--device",
-            self.cpu_device,
-            "--csv_column_names",
-            "Address",
-            "Tags",
-            "--csv_column_separator",  # Our dataset use a comma as separator
-            ",",
-        ]
+        parser_params = self.set_up_params(csv_column_names=["Address", "Tags"], csv_column_separator=",")
 
-        parser_params.extend(self.parser_test_default_settings)
         retrain.main(parser_params)
 
         self.assertTrue(
@@ -170,36 +159,18 @@ class RetrainTests(RetrainTestCase):
             )
         )
 
-    @skipIf(
-        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
-        "download of model too long for test in runner",
-    )
     def test_ifIsCSVFile_noColumnName_raiseValueError(self):
         with self.assertRaises(ValueError):
-            parser_params = [
-                self.a_fasttext_model_type,
-                self.a_train_csv_dataset_path,
-                "--device",
-                self.cpu_device,
-            ]
+            # We set up the params with the default value of csv_column_names of the test case method set_up_params,
+            # which is None, thus no column names.
+            parser_params = self.set_up_params(train_dataset_path=self.a_train_csv_dataset_path)
 
-            parser_params.extend(self.parser_test_default_settings)
             retrain.main(parser_params)
 
-    @skipIf(
-        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
-        "download of model too long for test in runner",
-    )
     def test_ifIsNotSupportedFile_raiseValueError(self):
         with self.assertRaises(ValueError):
-            parser_params = [
-                self.a_fasttext_model_type,
-                "an_unsupported_extension.json",
-                "--device",
-                self.cpu_device,
-            ]
+            parser_params = self.set_up_params(train_dataset_path="an_unsupported_extension.json")
 
-            parser_params.extend(self.parser_test_default_settings)
             retrain.main(parser_params)
 
     def test_givenSetOfParams_whenParseTrainingParams_thenReturnProperParams(self):
@@ -252,10 +223,6 @@ class RetrainTests(RetrainTestCase):
             if key in override_keys:
                 self.assertTrue(value in override_values)
 
-    @skipIf(
-        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
-        "download of model too long for test in runner",
-    )
     def test_integration_multiple_retrain_overrides(self):
         train_ratio = "0.8"
         batch_size = "64"
@@ -279,6 +246,16 @@ class RetrainTests(RetrainTestCase):
         retrain.main(parser_params)
 
         self.assertTrue(os.path.isfile(os.path.join(self.temp_checkpoints_obj.name, "checkpoints", "AName.ckpt")))
+
+    def test_ifCachePath_thenUseNewCachePath(self):
+        with patch("deepparse.cli.retrain.AddressParser") as address_parser_mock:
+            parser_params = self.set_up_params(cache_dir=self.a_cache_dir, epochs="1")
+            retrain.main(parser_params)
+
+            address_parser_mock.assert_called()
+            address_parser_mock.assert_called_with(
+                device=self.cpu_device, cache_dir=self.a_cache_dir, model_type=self.a_fasttext_model_type
+            )  # Default tests case default model type is the FastText model
 
 
 if __name__ == "__main__":